+++ ebzl/modules/bundle.py
@@ -17,7 +17,7 @@ def get_argument_parser():
     parameters.add_profile(parser)
     parameters.add_app_name(parser)
     parameters.add_region(parser, required=False)
-    parameters.add_version_label(parser)
+    parameters.add_version_label(parser, required=False)
 
     parser.add_argument("-d", "--docker-image",
                         required=True,