+++ ebzl/modules/create.py
@@ -15,7 +15,7 @@ def get_argument_parser():
 
     parameters.add_profile(parser)
     parameters.add_app_name(parser)
-    parameters.add_version_label(parser)
+    parameters.add_version_label(parser, required=False)
     parameters.add_region(parser, required=False)
     parameters.add_force(parser)