+++ ebzl/modules/deploy.py
@@ -11,7 +11,7 @@ def get_argument_parser():
     parser = argparse.ArgumentParser("ebzl deploy")
 
     parameters.add_profile(parser)
-    parameters.add_version_label(parser)
+    parameters.add_version_label(parser, required=False)
     parameters.add_region(parser, required=False)
 
     parser.add_argument("-e", "--env-name",